Remove ImapLibBasedProvider trait. Move all methods to ImapMail class. Add pecl install for mailparse extension. Some minor styles fixes. Enable ZendMailCest tesing.
