+++ .travis.yml
@@ -3,6 +3,8 @@ php:
   - 5.6
   - 5.5
 before_install:
+  - if [[ "$TRAVIS_PHP_VERSION" = "5.5" ]]; then pecl install mailparse-2.1.6 ; fi
+  - if [[ "$TRAVIS_PHP_VERSION" = "5.6" ]]; then pecl install mailparse-2.1.6 ; fi
   - phpenv config-add travis-php-config.ini
   - gem install mailcatcher
   - mailcatcher --http-ip 127.0.0.1 --http-port 1081 --smtp-ip 127.0.0.1 --smtp-port 1026
