+++ test/tests/_support/_generated/AcceptanceTesterActions.php
@@ -1,4 +1,4 @@
-<?php  //[STAMP] 9b2edb1cad29dc674cb7675512d47746
+<?php  //[STAMP] 6de008dbca8206fac00ca9b9544b7cfb
 namespace _generated;
 
 // This class was automatically generated by build task
